Création du component SelectIngredients pour simplication de FormCreateRecipe
