+++ frontend/src/pages/user/pages/CreateRecipe/components/formCreateRecipe/FormCreateRecipe.jsx
@@ -14,6 +14,7 @@ import SelectLevel from './components/select/selectLevel/SelectLevel';
 import InputTitle from './components/inputTitle/InputTitle';
 import InputImage from './components/inputImage/InputImage';
 import InputSteps from './components/inputSteps/InputSteps';
+import SelectIngredients from './components/select/selectIngredients/SelectIngredients';
 
 
 // Définir le schéma de validation avec yup
@@ -22,8 +23,8 @@ const FormCreateRecipe = () => {
   const { user } = useContext(UserContext);
   console.log("user in FormCreateRecipe :", user);
   const [ingredients, setIngredients] = useState([]);
-  const [steps, setSteps] = useState([]);
   const [selectedIngredients, setSelectedIngredients] = useState([]);
+  const [steps, setSteps] = useState([]);
   const [error, setError] = useState(null);
 
   // Initialiser les fonctions du formulaire avec react-hook-form
@@ -31,63 +32,15 @@ const FormCreateRecipe = () => {
     resolver: yupResolver(schemaValidationCreateRecipe),
   });
 
-  // Charger les ingrédients lors du premier rendu
-  useEffect(() => {
-    const fetchIngredients = async () => {
-      try {
-        const response = await getAllIngredients();
-        setIngredients(response);
-      } catch (err) {
-        console.error("Erreur lors de la récupération des ingrédients", err);
-        setError("Impossible de charger les ingrédients. Veuillez réessayer plus tard.");
-      }
-    };
-    fetchIngredients();
-  }, []);
 
   // Gérer l'ajout des ingrédients sélectionnés
-  const handleIngredientSelect = (event) => {
-    const selectedId = event.target.value;
-    const ingredient = ingredients.find(ing => ing._id === selectedId);
-    if (ingredient && !selectedIngredients.some(ing => ing._id === selectedId)) {
-      setSelectedIngredients([...selectedIngredients, ingredient]);
-    }
-  };
-
-  // Gérer la suppression des ingrédients sélectionnés
-  const removeIngredient = (id) => {
-    setSelectedIngredients(selectedIngredients.filter(ing => ing._id !== id));
-  };
 
   // Ajouter une étape à la liste des étapes
 
-  // Filtrer les options d'ingrédients en fonction de ceux qui sont déjà sélectionnés
-  const filteredIngredientOptions = ingredients
-    .filter(ing => !selectedIngredients.some(selected => selected._id === ing._id))
-    .map(ingredient => (
-      <option key={ingredient._id} value={ingredient._id}>
-        {ingredient.name}
-      </option>
-    ));
 
 
   // Convertir les étapes en éléments de liste
 
-  // Convertir les ingrédients sélectionnés en éléments d'affichage
-  const selectedIngredientsList = selectedIngredients.map(ingredient => (
-    <div key={ingredient._id} className="bg-white rounded-lg shadow p-4 flex flex-col items-center">
-      <img src={ingredient.image} alt={ingredient.name} className="w-24 h-24 object-cover rounded-full mb-2" />
-      <p className="text-center text-sm font-medium">{ingredient.name}</p>
-      <button
-        type="button"
-        onClick={() => removeIngredient(ingredient._id)}
-        className="mt-2 text-red-600 hover:text-red-800 text-sm flex items-center"
-      >
-        <FontAwesomeIcon icon={faMinus} className="mr-1" />
-        Retirer
-      </button>
-    </div>
-  ));
 
   // Gérer la soumission du formulaire
   const onSubmit = async (data) => {
@@ -150,23 +103,7 @@ const FormCreateRecipe = () => {
           <SelectLevel register={register} errors={errors} />
         </div>
 
-        <div>
-          <label htmlFor="ingredients" className="flex items-center text-sm font-medium text-orange-700">
-            <FontAwesomeIcon icon={faPlus} className="mr-2" />
-            Ajouter des ingrédients
-          </label>
-          <select
-            onChange={handleIngredientSelect}
-            className="mt-1 block w-full rounded-md border-orange-300 shadow-sm focus:border-orange-500 focus:ring focus:ring-orange-200 focus:ring-opacity-50"
-          >
-            <option value="">Sélectionnez un ingrédient</option>
-            {filteredIngredientOptions}
-          </select>
-        </div>
-
-        <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4">
-          {selectedIngredientsList}
-        </div>
+        <SelectIngredients register={register} errors={errors} ingredients={ingredients} setIngredients={setIngredients} selectedIngredients={selectedIngredients} setSelectedIngredients={setSelectedIngredients} />
           <InputSteps register={register} errors={errors} steps={steps} setSteps={setSteps} />
         <div>
           <button type="submit" className="w-full bg-orange-600 text-white px-4 py-2 rounded-md hover:bg-orange-700 focus:outline-none focus:ring-2 focus:ring-orange-500 focus:ring-opacity-50 transition duration-300 flex items-center justify-center">
